+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/reference/CDSPosition.java
@@ -12,9 +12,9 @@ import de.charite.compbio.jannovar.impl.util.StringUtil;
 public final class CDSPosition {
 
 	/** the transcript that this position is relative to */
-	public final TranscriptModel transcript;
+	private final TranscriptModel transcript;
 	/** the position within the transcript */
-	public final int pos;
+	private final int pos;
 
 	/** construct transcript position with one-based coordinate system */
 	public CDSPosition(TranscriptModel transcript, int pos) {
@@ -28,6 +28,16 @@ public final class CDSPosition {
 			this.pos = pos - 1;
 		else
 			this.pos = pos;
+	}
+
+	/** @return the transcript that this position is relative to */
+	public TranscriptModel getTranscript() {
+		return transcript;
+	}
+
+	/** @return the position within the transcript */
+	public int getPos() {
+		return pos;
 	}
 
 	/**